Everywhere: Make `Message` a sealed class
